Integrate matrix bot
* login matrix bot in main
* move github_bot::update to bots::update
* remove glitch-in-the-matrix
* remove tokio-core
