+++ src/lib.rs
@@ -1,16 +1,16 @@
+#![feature(result_map_or)]
+
 pub mod github_bot;
-pub mod issue;
 pub mod matrix_bot;
 pub mod pull_request;
-pub mod review_request;
 //pub mod team;
 //pub mod project;
+pub mod bots;
 pub mod db;
-pub mod user;
 pub mod error;
 pub mod github;
 pub mod matrix;
 pub mod repository;
-pub mod review;
+pub mod user;
 
 pub type Result<T> = std::result::Result<T, error::Error>;
