+++ src/user.rs
@@ -1,11 +1,11 @@
 use crate::{error, github, github_bot::GithubBot, Result};
 
 impl github::User {
-        pub fn riot_id(&self) -> String {
-                unimplemented!();
-        }
+	pub fn riot_id(&self) -> String {
+		unimplemented!();
+	}
 
 	pub fn is_core_developer(&self) -> bool {
-                unimplemented!();
-        }
+		unimplemented!();
+	}
 }
